refactor: move bindings to contracts mod
