+++ crates/contracts/Cargo.toml
@@ -1,0 +1,45 @@
+[package]
+name = "tempo-zone-contracts"
+description = "Tempo Zone contract bindings and ABI definitions"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# zones
+zone-primitives = { workspace = true, default-features = false }
+
+# alloy
+alloy-primitives = { workspace = true, default-features = false }
+alloy-sol-types = { workspace = true, default-features = false }
+alloy-contract = { workspace = true, optional = true }
+alloy-network = { workspace = true, optional = true }
+alloy-provider = { workspace = true, optional = true }
+futures = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true }
+serde = { workspace = true, optional = true }
+
+[dev-dependencies]
+const-hex.workspace = true
+
+[features]
+default = ["std", "serde", "rpc"]
+std = [
+    "alloy-primitives/std",
+    "alloy-sol-types/std",
+    "serde?/std",
+    "zone-primitives/std",
+]
+serde = ["dep:serde", "alloy-primitives/serde"]
+rpc = [
+    "dep:alloy-contract",
+    "dep:alloy-network",
+    "dep:alloy-provider",
+    "dep:futures",
+    "dep:tokio",
+]
