+++ crates/precompiles/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 
 [dependencies]
 # zones
+tempo-zone-contracts = { workspace = true, default-features = false }
 zone-primitives = { workspace = true, default-features = false }
 
 # tempo (workspace deps already have default-features = false)
@@ -46,6 +47,7 @@ tempo-precompiles = { workspace = true, features = ["test-utils"] }
 default = ["std"]
 test-utils = ["tempo-precompiles/test-utils"]
 std = [
+    "tempo-zone-contracts/std",
     "zone-primitives/std",
     "alloy-evm/std",
     "alloy-primitives/std",
