+++ crates/primitives/Cargo.toml
@@ -13,24 +13,13 @@ workspace = true
 [dependencies]
 alloy-primitives = { workspace = true }
 alloy-rlp = { workspace = true }
-alloy-sol-types = { workspace = true }
-alloy-contract = { workspace = true, optional = true }
-alloy-network = { workspace = true, optional = true }
-alloy-provider = { workspace = true, optional = true }
-futures = { workspace = true, optional = true }
-tokio = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 
-[dev-dependencies]
-const-hex.workspace = true
-
 [features]
-default = ["std", "serde", "rpc"]
+default = ["std", "serde"]
 std = [
     "alloy-primitives/std",
     "alloy-rlp/std",
-    "alloy-sol-types/std",
     "serde?/std",
 ]
 serde = ["dep:serde", "alloy-primitives/serde"]
-rpc = ["dep:alloy-contract", "dep:alloy-network", "dep:alloy-provider", "dep:futures", "dep:tokio"]
