+++ crates/tempo-zone/Cargo.toml
@@ -22,8 +22,9 @@ tempo-transaction-pool.workspace = true
 tempo-alloy.workspace = true
 tempo-contracts.workspace = true
 tempo-precompiles.workspace = true
+tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
 zone-precompiles = { workspace = true, features = ["std"] }
-zone-primitives = { workspace = true, features = ["std", "serde", "rpc"] }
+zone-primitives = { workspace = true, features = ["std", "serde"] }
 zone-rpc.workspace = true
 
 # reth
